Bug 828088 - Rework richgrid and richgriditem bindings to use css columns for down-then-across grids. Hardening downloads and tests. r=fryn

+++ browser/metro/base/content/downloads.js
@@ -20,8 +20,14 @@ var Downloads = {
 
   _getLocalFile: function dh__getLocalFile(aFileURI) {
     // XXX it's possible that using a null char-set here is bad
-
-    const fileUrl = Services.io.newURI(aFileURI.spec || aFileURI, null, null).QueryInterface(Ci.nsIFileURL);
+    let spec = ('string' == typeof aFileURI) ? aFileURI : aFileURI.spec;
+    let fileUrl;
+    try {
+      fileUrl = Services.io.newURI(spec, null, null).QueryInterface(Ci.nsIFileURL);
+    } catch (ex) {
+      Util.dumpLn("_getLocalFile: Caught exception creating newURI from file spec: "+aFileURI.spec+": " + ex.message);
+      return;
+    }
     return fileUrl.file.clone().QueryInterface(Ci.nsILocalFile);
   },
 
@@ -43,13 +49,21 @@ var Downloads = {
   openDownload: function dh_openDownload(aDownload) {
     // expects xul item
     let id = aDownload.getAttribute("downloadId");
-    let download = this.manager.getDownload(id)
+    let download = this.manager.getDownload(id);
     let fileURI = download.target;
-    let file = this._getLocalFile(fileURI);
 
+    if (!(fileURI && fileURI.spec)) {
+      Util.dumpLn("Cant open download "+id+", fileURI is invalid");
+      return;
+    }
+
+    let file = this._getLocalFile(fileURI);
     try {
-      file.launch();
-    } catch (ex) { }
+      file && file.launch();
+    } catch (ex) {
+      Util.dumpLn("Failed to open download, with id: "+id+", download target URI spec: " + fileURI.spec);
+      Util.dumpLn("Failed download source:"+(aDownload.source && aDownload.source.spec));
+    }
   },
 
   removeDownload: function dh_removeDownload(aDownload) {
@@ -66,14 +80,24 @@ var Downloads = {
 
   cancelDownload: function dh_cancelDownload(aDownload) {
     let id = aDownload.getAttribute("downloadId");
-    let download = this.manager.getDownload(id)
+    let download = this.manager.getDownload(id);
     this.manager.cancelDownload(id);
 
     let fileURI = download.target;
-    let file = this._getLocalFile(fileURI);
 
-    if (file.exists())
-      file.remove(false);
+    if (!(fileURI && fileURI.spec)) {
+      Util.dumpLn("Cant remove download file for: "+id+", fileURI is invalid");
+      return;
+    }
+
+    let file = this._getLocalFile(fileURI);
+    try {
+      if (file && file.exists())
+        file.remove(false);
+    } catch (ex) {
+      Util.dumpLn("Failed to cancel download, with id: "+id+", download target URI spec: " + fileURI.spec);
+      Util.dumpLn("Failed download source:"+(aDownload.source && aDownload.source.spec));
+    }
   },
 
   pauseDownload: function dh_pauseDownload(aDownload) {
@@ -93,7 +117,7 @@ var Downloads = {
 
   showPage: function dh_showPage(aDownload) {
     let id = aDownload.getAttribute("downloadId");
-    let download = this.manager.getDownload(id)
+    let download = this.manager.getDownload(id);
     let uri = this._getReferrerOrSource(download);
     if (uri)
       BrowserUI.newTab(uri, Browser.selectedTab);
@@ -197,13 +221,14 @@ DownloadsView.prototype = {
   },
 
   _getAttrsForDownload: function dv__getAttrsForDownload(aDownload) {
-    // expects a DownloadManager download object
+    // params: nsiDownload
     return {
       typeName: 'download',
       downloadId: aDownload.id,
       downloadGuid: aDownload.guid,
       name: aDownload.displayName,
-      target: aDownload.target,
+      // use the stringified version of the target nsIURI for the item attribute
+      target: aDownload.target.spec,
       iconURI: "moz-icon://" + aDownload.displayName + "?size=64",
       date: DownloadUtils.getReadableDates(new Date())[0],
       domain: DownloadUtils.getURIHost(aDownload.source.spec)[0],
@@ -215,6 +240,8 @@ DownloadsView.prototype = {
   _updateItemWithAttrs: function dv__updateItemWithAttrs(anItem, aAttrs) {
     for (let name in aAttrs)
       anItem.setAttribute(name, aAttrs[name]);
+    if (anItem.refresh)
+      anItem.refresh();
   },
 
   _getDownloadSize: function dv__getDownloadSize (aSize) {
@@ -322,8 +349,7 @@ DownloadsView.prototype = {
   },
 
   clearDownloads: function dv_clearDownloads() {
-    while (this._set.itemCount > 0)
-      this._set.removeItemAt(0);
+    this._set.clearAll();
   },
 
   addDownload: function dv_addDownload(aDownload) {
